% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageForecastPRO.R
\name{ageForecastPRO}
\alias{ageForecastPRO}
\title{ageForecastPRO Produces a Age Forecast data set and chart}
\usage{
ageForecastPRO(fips, stYr, mYr, eYr, base = 12, agegroup = "ten")
}
\arguments{
\item{fips}{is the fips code for the county being examined}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\value{
ggplot2 graphic and data file
}
\description{
Copied from "ms_popage" in codemgprofile, Modified by AB 12/2017
Creates a Chart comparing Forecast Population Growth by Age in Colorado.
}
\details{
Uses the data from the State Demography Office package codemog to
create a graph showing projected population  changes by Age for each Colorado county from
2010 to 2025.
The chart is modified from the original.  Now, we show three bars, one for each series.
}
