% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseIndustries.R
\name{baseIndustries}
\alias{baseIndustries}
\title{baseIndustries outputs the base_industries job plot and data}
\usage{
baseIndustries(fips, ctyname, curyr, oType, base = 10)
}
\arguments{
\item{fips}{numeric, county-level FIPS code}

\item{ctyname}{County Name string, from input$unit}

\item{curyr}{the current year}
}
\value{
ggplot2 graphic and data file
}
\description{
This includes code to output data for the Denver-Boulder MSA when Adams, Arapahoe, Boulder,
    Broomfield, Denver, Douglas, or Jefferson County are selected.
}
