% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxContent.R
\name{boxContent}
\alias{boxContent}
\title{boxContent outputs the HTML code for content and buttons of the various info boxes}
\usage{
boxContent(title, description, source, MSA, stats, table, urlList)
}
\arguments{
\item{title}{The title of the box content}

\item{description}{The long description of the Box}

\item{MSA}{Test describing the MSA adjustment for counties in the Denver MSA}

\item{stats}{A T/F value to output information about statistical tests}

\item{table}{A T/F value to outout information about the output table}

\item{urlList}{A list of output links to sources c(name, URL)}
}
\value{
Content and buttons for a specified info box,
}
\description{
need  to put the box text in a HTML() call to render
}
