% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captionSrc.R
\name{captionSrc}
\alias{captionSrc}
\title{captionSrc formats the captions for the tables and plots}
\usage{
captionSrc(type, dataSrc)
}
\arguments{
\item{type}{the data type, "SDO" for State Demography Office data, "ACS" for American Community Survey data}

\item{dataSrc}{The data string for ACS data}
}
\value{
Date-stamped caption string
}
\description{
captionSrc formats the captions for the tables and plots
}
