% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocPlot.R
\name{cocPlot}
\alias{cocPlot}
\title{cocPlot: Components of Change Chart, this is a county-level chart, regardless of output level}
\usage{
cocPlot(fips, ctyname, lYr, base = 12)
}
\arguments{
\item{fips}{numeric, county-level FIPS code  ** will need to extract county level FIPS code in all calls}

\item{ctyname}{County Name string, from input$unit}

\item{lYr}{the last year of the output date range}
}
\value{
ggplot2 graphic and data file
}
\description{
cocPlot: Components of Change Chart, this is a county-level chart, regardless of output level
}
