% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_timeseries.R
\name{county_timeseries}
\alias{county_timeseries}
\title{county_timeseries Creates a \code{ggplot2} chart of the population for a CO county}
\usage{
county_timeseries(fips, beginyear = 1990, endYear, base = 10)
}
\arguments{
\item{fips}{The County FIPS number (without leading Zeros)}

\item{beginyear}{The first year in the timeseries Defaults to 1990.}

\item{endYear}{The last year in the timeseries Defaults to 2013.}

\item{base}{Base font size.}
}
\value{
ggplot2 graphic and data file
}
\description{
This is a replacement for county_ts_chart  Copied from codemogProfile
}
\details{
Takes some basic input on the time period and county then creates a
plot of the data in \code{ggplot2}.  Similar to the county_ts_data()
function.  Can create timeseries from 1990 to 2040 (beyond 2013 are
forecasts).
}
