% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadObj.R
\name{downloadObj}
\alias{downloadObj}
\title{downloadObj  File Download Modules}
\usage{
downloadObj(input, output, session, place, oname, dobj)
}
\arguments{
\item{place}{is the place name, typically the value of input$unit}

\item{oname}{the input data description object}

\item{dboj}{is the data object to be output}
}
\description{
downloadObj is the server function that facilitates the download
}
