% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadObjUI.R
\name{downloadObjUI}
\alias{downloadObjUI}
\title{downloadObjUI  File Download Modules}
\usage{
downloadObjUI(id)
}
\arguments{
\item{id}{is the data name and creates the module/namespace ID}
}
\value{
output button type (plot or data)
}
\description{
downloadObjUI is the UI function that creates the download buttons
}
