% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educPRO.R
\name{educPRO}
\alias{educPRO}
\title{educPRO Creates a Chart comparing educational attainment of two areas}
\usage{
educPRO(fips, ctyname, state = "08", fips2 = "", state2 = "08", ACS,
  base = 10)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{state}{is the state that the original fips}

\item{fips2}{is the second area to be compared Defaults to Blank}

\item{state2}{is the state of the second place to be compared, is set to call up CO since fips2 is blank Defaults to 08}

\item{ACS}{Specifies the ACS data set to be used, reads curACS from Shiny program}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\value{
ggplot2 graphic aand data file
}
\description{
Modified from ms_ed in codemogProfile AB 12/2017
Uses the codemog_api function to access ACS data (defaults to 13-5yr) to create a ggplot2 chart for
use in profiles.
}
