% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateVenn.R
\name{GenerateVenn}
\alias{GenerateVenn}
\title{GenerateVenn Generates a Venn diagram using LODES data}
\usage{
GenerateVenn(fips, ctyname, oType)
}
\arguments{
\item{fips}{is the numeric fips code for county}

\item{ctyname}{is the cplace name from input$unit}

\item{oType}{output type html or latex}
}
\value{
ggplot2 graphic, formatted datatables, and datasets
}
\description{
GenerateVenn Generates a Venn diagram using LODES data
}
