% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseEstPRO.R
\name{houseEstPRO}
\alias{houseEstPRO}
\title{houseEstPRO Produces a plot of the housing projections from 2010 to 2050
  from the household_projections data table}
\usage{
houseEstPRO(fips, ctyname, curYr, base = 10)
}
\arguments{
\item{fips}{is the numeric fips code for county}

\item{ctyname}{is the cplace name from input$unit}

\item{curYr}{Ins the current year value}

\item{base}{is the abse text size for the ggplot2 object and codemog_theme()}
}
\value{
ggplot2 graphic and data file
}
\description{
houseEstPRO Produces a plot of the housing projections from 2010 to 2050
  from the household_projections data table
}
