% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housePRO.R
\name{housePRO}
\alias{housePRO}
\title{housePRO  Produces the housing table
 CO Housing Unit Table}
\usage{
housePRO(fips, ctyname, ACS, oType, state = "08")
}
\arguments{
\item{fips}{The FIPS of the Place or County to use for the graph}

\item{ctyname}{The place Name}

\item{ACS}{The American Community Survey Vintage}

\item{oType}{Output type, html or latex}

\item{state}{Defaults to Colorado}
}
\value{
kable formatted table and data file
}
\description{
This function compares housing occupancy and vacancy rates for a place to the state
}
