% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HouseVal.R
\name{HouseVal}
\alias{HouseVal}
\title{HouseVal Table Showing the Median House value for Owner-occupied housing,
    the Median Gross Rent, and Median Costs as a Percentage of Income for
    Owners and Renters for a place and the State of Colorado}
\usage{
HouseVal(fips, ctyname, ACS, oType, state = "08")
}
\arguments{
\item{fips}{The FIPS of the Place or County to use for the graph}

\item{ctyname}{The place Name}

\item{ACS}{The American Community Survey Vintage}

\item{oType}{Output Type, html or latex}

\item{state}{defaults to Colorado}
}
\value{
kable formatted  table and data file
}
\description{
HouseVal Table Showing the Median House value for Owner-occupied housing,
    the Median Gross Rent, and Median Costs as a Percentage of Income for
    Owners and Renters for a place and the State of Colorado
}
