% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncomePRO.R
\name{incomePRO}
\alias{incomePRO}
\title{incomePRO Income Distribution Graph}
\usage{
incomePRO(fips, ctyname, fips2 = "", state = "08", state2 = "08",
  ACS = "acs1115", base = 12)
}
\arguments{
\item{fips}{The FIPS of the Place or County to use}

\item{fips2}{The FIPS of the Place or County to use for comparison}

\item{state}{The State FIPS to use.  Defaults to CO.}

\item{state2}{The State FIPS to use as comparison.  Defaults to CO.}

\item{ACS}{Data set id, eg: for 2015 ACS: "acs1115"}
}
\value{
ggplot2 graphic and data file
}
\description{
Modified from ms_income in codemogProfile AB 12/2017
}
\details{
This function pulls data and generates a graph of the income distribution
 for the areas selected based on the ACS
}
