% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobMigration.R
\name{jobMigration}
\alias{jobMigration}
\title{jobMigration Creates a Chart showing the number of jobs and net migration trends
 1985 to present}
\usage{
jobMigration(fips, ctyname, maxyr, base = 10)
}
\arguments{
\item{fips}{is the fips code for the county being examined}

\item{ctyname}{This parameter puts the name of the county in the chart}

\item{maxyr}{The maximum year value, from CurYr}

\item{base}{is the abse text size for the ggplot2 object and codemog_theme(), defaults to base = 10}
}
\value{
ggplot graphic and data file
}
\description{
Uses State Demography Office data to create a chart comparing jobs by major sector
for a selected place, compared to Colorado.
}
\details{
Changes format to a basic bar chart with text annotations from the original formattable output.
}
