% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobsByIndustry.R
\name{jobsByIndustry}
\alias{jobsByIndustry}
\title{jobsByIndustry Creates a Chart showing the Total estimated jobs by sector for
  a place and for Colorado}
\usage{
jobsByIndustry(fips, ctyname, curyr, base = 10)
}
\arguments{
\item{fips}{is the fips code for the county being examined}

\item{ctyname}{This parameter puts the name of the county in the chart}

\item{curyr}{The maximum year value, from CurYr}

\item{base}{is the abse text size for the ggplot2 object and codemog_theme(), defaults to base = 10}
}
\value{
ggplot graphic and data file
}
\description{
Uses State Demography Office data to create a chart compaing jobs by major sector
for a selected place, compared to Colorado.
}
