% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobsPopForecast.R
\name{jobsPopForecast}
\alias{jobsPopForecast}
\title{jobsPopForecast Produces a plot and table showing changes in forecasted jobs and population
 for the period from 2010 to 2035}
\usage{
jobsPopForecast(fips, ctyname, base = 10)
}
\arguments{
\item{fips}{is the fips code for the county being examined}

\item{ctyname}{the name of the place being output}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\value{
ggplot2 graphic, a html or latex table and a dataset
}
\description{
This includes code to output data for the Denver-Boulder MSA when Adams, Arapahoe, Boulder,
    Broomfield, Denver, Douglas, or Jefferson County are selected.
}
