% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTofips.R
\name{listTofips}
\alias{listTofips}
\title{listTofips : Produces a vector of FIPS codes from an inpout list of Census County and Plance Name Codes.}
\usage{
listTofips(df, level, inList1)
}
\arguments{
\item{df}{The census place look up file, produced by popPlace.}

\item{level}{identifies the level to be used (State, Plannign Regions, Counties, Municipalities/Places)
taken from the input$level parameter from the dashboard}

\item{inList1}{The list of place names,  Comes from input$unit or input$comp2.}
}
\value{
the fipscode(s) for a selected data level
}
\description{
listTofips : Produces a vector of FIPS codes from an inpout list of Census County and Plance Name Codes.
}
