% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianAgeTab.R
\name{medianAgeTab}
\alias{medianAgeTab}
\title{medianAgeTab Creates table showing the Median Age by Gender
for a selected place and for the state}
\usage{
medianAgeTab(fips, ACS, ctyname, oType, state = "08")
}
\arguments{
\item{fips}{The County FIPS number (without leading Zeros)}

\item{ACS}{a string identifying the input dataset eg: "acs1115"}

\item{ctyname}{a string identiying the place name}

\item{oType}{type of output, html or latex}

\item{state}{the State FIPS code, defaluts to "08" for Colorado.}
}
\value{
a formatted table and dataset
}
\description{
medianAgeTab Creates table showing the Median Age by Gender
for a selected place and for the state
}
