% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migbyagePRO.R
\name{migbyagePRO}
\alias{migbyagePRO}
\title{migbyagePRO Creates a Chart showing the 2000-2010 net Migration rate by age}
\usage{
migbyagePRO(fips, ctyname, base = 10)
}
\arguments{
\item{fips}{is the numeric fips code for the main area to be compared}

\item{ctyname}{is the cplace name from input$unit}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\value{
ggplot2 graphic and data file
}
\description{
migbyagePRO Creates a Chart showing the 2000-2010 net Migration rate by age
}
