% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOHouse.R
\name{OOHouse}
\alias{OOHouse}
\title{OOHouse Summary table for owner-occupied Housing}
\usage{
OOHouse(fips, ctyname, ACS, oType, state = "08")
}
\arguments{
\item{fips}{The FIPS of the Place or County to use for the graph}

\item{ctyname}{The place Name}

\item{ACS}{The American Community Survey Vintage}

\item{oType}{Output Type, html or latex}

\item{state}{defaults to Colorado}
}
\value{
kable formatted  table and data file
}
\description{
OOHouse Summary table for owner-occupied Housing
}
