% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{percent: utility function  that ourputs numeric value (typically * 100) as a percentage}
\usage{
percent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{is the input numeric value.  The percentage value (y/total *100) is claculated outside of the function}

\item{digits}{the number of digits to output}

\item{format}{input format}
}
\value{
a numeric string formatted as a percentage
}
\description{
Maintains leading and trailing zeros...
  Taken from StackOverFlow 20171122...
}
