% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popForecast.R
\name{popForecast}
\alias{popForecast}
\title{popForecast Creates a Chart showing population and estmates}
\usage{
popForecast(fips, ctyname, byr = 1990, eyr = 2050, base = 10)
}
\arguments{
\item{fips}{is the numeric fips code for the main area to be compared}

\item{ctyname}{is the cplace name from input$unit}

\item{byr}{is the first year of the series to be extracted by county_sya (min 1990)}

\item{eyr}{is the last  year of the series to be extracted by county_sya (max 2050)}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\value{
ggplot2 graphic and data file
}
\description{
popForecast Creates a Chart showing population and estmates
}
