% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popPlace.R
\name{popPlace}
\alias{popPlace}
\title{popPlace : Populates the input$unit field using information from the PostGres estimates database.
 return a data frame with the placefips, countyfips, placename and totalPopulation}
\usage{
popPlace(level)
}
\arguments{
\item{level}{identifies the level to be used (State, Planning Regions, Counties, Municipalities/Places)
taken from the input$level parameter from the dashboard}
}
\description{
popPlace : Populates the input$unit field using information from the PostGres estimates database.
 return a data frame with the placefips, countyfips, placename and totalPopulation
}
