% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popTable.R
\name{popTable}
\alias{popTable}
\title{popTable The population table showing the annual growth rate in the Population Section}
\usage{
popTable(cty, ctyname, sYr, eYr, oType)
}
\arguments{
\item{cty}{short FIPS code, without the state code}

\item{ctyname}{the place name}

\item{sYr}{Start Year}

\item{eYr}{End year}

\item{oType}{Output Type, html or latex}
}
\value{
kable formatted  table and data file
}
\description{
popTable The population table showing the annual growth rate in the Population Section
}
