% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceTab1.R
\name{raceTab1}
\alias{raceTab1}
\title{raceTab1 Table showing the percentage values by ethnic/race categories}
\usage{
raceTab1(fips, ctyname, ACS, oType)
}
\arguments{
\item{fips}{the short FIPS code}

\item{ctyname}{Place Name}

\item{ACS}{data depository from the American Community Survey API}

\item{oType}{Type of Output html, latex}
}
\value{
kable formatted  table and data file
}
\description{
pulls data from API This table shows a set of histoical comparisons between
   the 2000 Census, the 2010 Census and the latest ACS API
}
\details{
This table does not report MOEs for ACS series, because of the lack of cunsus MOEs...
}
