% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceTab2.R
\name{raceTab2}
\alias{raceTab2}
\title{raceTab2 Table showing the percentage values by ethnic/race categories}
\usage{
raceTab2(fips, ctyname, ACS, oType)
}
\arguments{
\item{fips}{the short FIPS code}

\item{ctyname}{Place Name}

\item{ACS}{data depository from the American Community Survey API}

\item{oType}{Type of output, html or latex}
}
\value{
kable formatted table and data file
}
\description{
pulls data from API This table compares Colorado % to selected geography
}
\details{
This table reports the MOEs and a significance test for each series
   comparing the percentages from each table...
}
