% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residentialLF.R
\name{residentialLF}
\alias{residentialLF}
\title{residentialLF Creates a two plots showing residential labor force participation}
\usage{
residentialLF(fips, ctyname, base = 10)
}
\arguments{
\item{fips}{is the numeric fips code for the main area to be compared}

\item{ctyname}{is the cplace name from input$unit}

\item{base}{is the abse text size for the ggplot2 object and codemog_theme()}
}
\value{
ggplot2 graphics and associated data sets
}
\description{
Plot 1: a line chart showing forcasted trend line for persons in the labor force and
    persons age 16 and older.
Plor2: a bar chart comparing forcasted labor force participation for a place and Colorado
}
