% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundUpNice.R
\name{roundUpNice}
\alias{roundUpNice}
\alias{roundUpNice}
\title{roundUpNice: rounds numbers to celing or floor based on inputs  used in setting the range for the COC chart}
\usage{
roundUpNice(x, Unit)

roundUpNice(x, Unit)
}
\arguments{
\item{x}{is the input numeric value.}

\item{Unit}{the base value to round around}
}
\value{
numeric value rounded up to the specific number of digits
}
\description{
Maintains leading and trailing zeros...
  Taken from StackOverFlow 20171122...

roundUpNice: rounds numbers to celing or floor based on inputs  used in setting the range for the COC chart
  Maintains leading and trailing zeros...
  Taken from StackOverFlow 20171122...
  @param x is the input numeric value.
  @param Unit the base value to round around
  @return numeric value rounded up to the specific number of digits
  @export
}
