% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTHouse.R
\name{RTHouse}
\alias{RTHouse}
\title{RTHouse Summary table for Rental Housing}
\usage{
RTHouse(fips, ctyname, ACS, oType, state = "08")
}
\arguments{
\item{fips}{The FIPS of the Place or County to use for the graph}

\item{ctyname}{The place Name}

\item{ACS}{The American Community Survey Vintage}

\item{oType}{Output Type html or latex}

\item{state}{defaults to Colorado}
}
\value{
kable formatted table and data file
}
\description{
RTHouse Summary table for Rental Housing
}
