% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setYrRange.R
\name{setYrRange}
\alias{setYrRange}
\title{setYrRange Calculates a range of year values, typically in 5-year increments with adjustments
  accounting for StartYr and EndYr values that are not multiples of 5}
\usage{
setYrRange(StartYr, EndYr)
}
\arguments{
\item{StartYr}{a numberic variable for the start of the series}

\item{EndYr}{a numeric variable for the end of the series}
}
\value{
a numeric vector of the years between StartYr and EndYr accounting for odd years
}
\description{
setYrRange Calculates a range of year values, typically in 5-year increments with adjustments
  accounting for StartYr and EndYr values that are not multiples of 5
}
