% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{simpleCap convers strings to proper case,}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{input string}
}
\value{
String formatted in Proper Case
}
\description{
stolen from Stackoverflow.
}
