% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabList.R
\name{tabList}
\alias{tabList}
\title{tabList returns the list items from the outputList  based on the value of input$outChk}
\usage{
tabList(item)
}
\arguments{
\item{item}{the item name in input$outChk}
}
\value{
the content of each panel, drawn ffrom the output lists defined in the
     GLOBAL section of the  UI code
}
\description{
tabList returns the list items from the outputList  based on the value of input$outChk
}
