% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabTitle.R
\name{tabTitle}
\alias{tabTitle}
\title{tabTitle manages the output of descriptive tabs in the interface}
\usage{
tabTitle(item)
}
\arguments{
\item{item}{the item name in input$outChk}
}
\value{
Descriptive string provided in the tabs of the main interface
}
\description{
tabTitle manages the output of descriptive tabs in the interface
}
