% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyWages.R
\name{weeklyWages}
\alias{weeklyWages}
\title{weeklyWages Produces a plot and dataset showing  average weekly wages
 for the period from 2010 to the present}
\usage{
weeklyWages(fips, ctyname, base = 10)
}
\arguments{
\item{fips}{is the fips code for the county being examined}

\item{ctyname}{the name of the place being output}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\value{
ggplot2 graphic, a html or lates table and a dataset
}
\description{
weeklyWages Produces a plot and dataset showing  average weekly wages
 for the period from 2010 to the present
}
